LC-single-number-iii [Runtime: N/A; Memory: N/A]

+++ bit-manipulation/single-number-iii.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
     vector<int> singleNumber(vector<int>& nums) {
-        int x=0;
+        long long x=0;
         for(int i=0;i<nums.size();i++) x=x^nums[i];
 
-        int rightmost=x&-x;
+        long long rightmost= x & -x;
         int b2=0,b1=0;
         for(int i=0;i<nums.size();i++){
             if(nums[i]&rightmost)
